[Vuepress] Add basic visit counter
The code is really basic: async request with no content to server
containing a counter increased on each request, with the value
logged once every day to a CSV file to load it into Grafana.

The server is designed to be as privacy-friendly as possible, with
no HTTP logs kept. See code at https://codeberg.org/Yaya-Cout/HTTP-Counter.

A future update may add rate-limiting on the reverse proxy side to
count unique visitors.